@@ -6,7 +6,7 @@
     ],
     "license": "MIT",
     "require": {
-        "php": ">=5.5"
+        "php": "^8.3 || ^8.4"
     },
     "autoload": {
         "psr-4": {
@@ -14,8 +14,7 @@
         }
     },
     "require-dev": {
-        "phpunit/phpunit": "4.8 - 5.7",
-        "gamegos/php-code-sniffer": "*"
+        "phpunit/phpunit": "^10.0 || ^11.0"
     },
     "autoload-dev": {
         "psr-4": {
